Use subString to split the band name
Use subString to split the band name and try to extract the method from the class.
If the last char is the same as the first letter, repet the char from second to last one.
If the last char is different from the first one, add The before the str.